Split scenario configs and parser into different modules
- The scenario examples (xml, xosc) can now be found in srunner/examples
- The scenario configuration modules were split into the configuration
itself (in srunner/scenarioconfigs) and the parsers (in srunner/tools)
