+++ .travis.yml
@@ -24,6 +24,7 @@ matrix:
         - autopep8 srunner/scenariomanager/scenarioatomics/*.py --in-place --max-line-length=120
         - autopep8 srunner/scenarios/*.py --in-place --max-line-length=120
         - autopep8 srunner/tools/*.py --in-place --max-line-length=120
+        - autopep8 srunner/scenarioconfigs/*.py --in-place --max-line-length=120
         - autopep8 scenario_runner.py --in-place --max-line-length=120
         - git diff --quiet HEAD --; if [ ! $? -eq 0 ]; then echo "Code is not autopep8 compliant. Please run code_check_and_formatting.sh"; git diff HEAD --; exit 1; fi
       after_failure:
@@ -43,10 +44,12 @@ matrix:
         - pylint --rcfile=.pylintrc --disable=I srunner/scenariomanager
         - pylint --rcfile=.pylintrc --disable=I srunner/scenarios
         - pylint --rcfile=.pylintrc --disable=I srunner/tools
+        - pylint --rcfile=.pylintrc --disable=I srunner/scenarioconfigs
         - pylint --rcfile=.pylintrc --disable=I scenario_runner.py
         - score=`pylint --rcfile=.pylintrc --disable=I srunner/scenariomanager | grep -i "rated at" | awk '{print $7}'`; if [ "$score" != "10.00/10" ]; then static_code_quality_passed=0; fi
         - score=`pylint --rcfile=.pylintrc --disable=I srunner/scenarios | grep -i "rated at" | awk '{print $7}'`; if [ "$score" != "10.00/10" ]; then static_code_quality_passed=0; fi
         - score=`pylint --rcfile=.pylintrc --disable=I srunner/tools | grep -i "rated at" | awk '{print $7}'`; if [ "$score" != "10.00/10" ]; then static_code_quality_passed=0; fi
+        - score=`pylint --rcfile=.pylintrc --disable=I srunner/scenarioconfigs | grep -i "rated at" | awk '{print $7}'`; if [ "$score" != "10.00/10" ]; then static_code_quality_passed=0; fi
         - score=`pylint --rcfile=.pylintrc --disable=I scenario_runner.py | grep -i "rated at" | awk '{print $7}'`; if [ "$score" != "10.00/10" ]; then static_code_quality_passed=0; fi
         - if [ $static_code_quality_passed -eq 0 ]; then echo "Code is not pylint compliant. Please run code_check_and_formatting.sh"; exit 1; fi
       after_failure:
