+++ srunner/examples/BackgroundActivity.xml
