+++ srunner/examples/ControlLoss.xml
