+++ srunner/examples/CyclistCrossing.xosc
