+++ srunner/examples/NoSignalJunction.xml
