+++ srunner/examples/ObjectCrossing.xml
