+++ srunner/examples/OppositeDirection.xml
