+++ srunner/examples/RunningRedLight.xml
