+++ srunner/examples/SignalizedJunctionLeftTurn.xml
