+++ srunner/examples/SignalizedJunctionRightTurn.xml
