+++ srunner/examples/signalized_junction_left_turn.xml
