+++ srunner/scenarios/master_scenario.py
@@ -10,7 +10,7 @@ Basic CARLA Autonomous Driving training scenario
 
 import py_trees
 
-from srunner.tools.config_parser import RouteConfiguration
+from srunner.scenarioconfigs.route_scenario_configuration import RouteConfiguration
 from srunner.scenariomanager.scenarioatomics.atomic_behaviors import Idle
 from srunner.scenariomanager.scenarioatomics.atomic_criteria import *
 from srunner.scenarios.basic_scenario import BasicScenario
