+++ srunner/tools/scenario_config_parser.py
@@ -13,137 +13,8 @@ import glob
 import os
 import xml.etree.ElementTree as ET
 
-import carla
-from agents.navigation.local_planner import RoadOption
-
-
-class RouteConfiguration(object):
-
-    """
-    This class provides the basic  configuration for a route
-    """
-
-    def __init__(self, route=None):
-        if route is not None:
-            self.data = route
-        else:
-            self.data = None
-
-    def parse_xml(self, node):
-        """
-        Parse route config XML
-        """
-        self.data = []
-
-        for waypoint in node.iter("waypoint"):
-            x = float(waypoint.attrib.get('x', 0))
-            y = float(waypoint.attrib.get('y', 0))
-            z = float(waypoint.attrib.get('z', 0))
-            c = waypoint.attrib.get('connection', '')
-            connection = RoadOption[c.split('.')[1]]
-
-            self.data.append((carla.Location(x, y, z), connection))
-
-
-class TargetConfiguration(object):
-
-    """
-    This class provides the basic  configuration for a target location
-    """
-
-    transform = None
-
-    def __init__(self, node):
-        pos_x = float(node.attrib.get('x', 0))
-        pos_y = float(node.attrib.get('y', 0))
-        pos_z = float(node.attrib.get('z', 0))
-
-        self.transform = carla.Transform(carla.Location(x=pos_x, y=pos_y, z=pos_z))
-
-
-class ActorConfigurationData(object):
-
-    """
-    This is a configuration base class to hold model and transform attributes
-    """
-
-    def __init__(self, model, transform, rolename='other', autopilot=False, random=False, amount=1):
-        self.model = model
-        self.rolename = rolename
-        self.transform = transform
-        self.autopilot = autopilot
-        self.random_location = random
-        self.amount = amount
-
-
-class ActorConfiguration(ActorConfigurationData):
-
-    """
-    This class provides the basic actor configuration for a
-    scenario:
-    - Location and rotation (transform)
-    - Model (e.g. Lincoln MKZ2017)
-    """
-
-    def __init__(self, node, rolename):
-
-        pos_x = float(node.attrib.get('x', 0))
-        pos_y = float(node.attrib.get('y', 0))
-        pos_z = float(node.attrib.get('z', 0))
-        yaw = float(node.attrib.get('yaw', 0))
-
-        random_location = False
-        if 'random_location' in node.keys():
-            random_location = True
-
-        autopilot = False
-        if 'autopilot' in node.keys():
-            autopilot = True
-
-        amount = 1
-        if 'amount' in node.keys():
-            amount = int(node.attrib['amount'])
-
-        super(ActorConfiguration, self).__init__(node.attrib.get('model', 'vehicle.*'),
-                                                 carla.Transform(carla.Location(x=pos_x, y=pos_y, z=pos_z),
-                                                 carla.Rotation(yaw=yaw)),
-                                                 node.attrib.get('rolename', rolename),
-                                                 autopilot, random_location, amount)
-
-
-class WeatherConfiguration(object):
-
-    """
-    This class provides basic weather configuration values
-    """
-
-    cloudyness = -1
-    precipitation = -1
-    precipitation_deposits = -1
-    wind_intensity = -1
-    sun_azimuth = -1
-    sun_altitude = -1
-
-
-class ScenarioConfiguration(object):
-
-    """
-    This class provides a basic scenario configuration incl.:
-    - configurations for all actors
-    - town, where the scenario should be executed
-    - name of the scenario (e.g. ControlLoss_1)
-    - type is the class of scenario (e.g. ControlLoss)
-    """
-
-    trigger_points = []
-    ego_vehicles = []
-    other_actors = []
-    town = None
-    name = None
-    type = None
-    target = None
-    route = None
-    weather = WeatherConfiguration()
+from srunner.scenarioconfigs.scenario_configuration import ScenarioConfiguration, ActorConfiguration
+from srunner.scenarioconfigs.route_scenario_configuration import RouteConfiguration, TargetConfiguration
 
 
 class ScenarioConfigurationParser(object):
@@ -219,8 +90,8 @@ class ScenarioConfigurationParser(object):
         Parse *all* config files and provide a list with all scenarios @return
         """
 
-        list_of_config_files = glob.glob("{}/srunner/configs/*.xml".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
-        list_of_config_files += glob.glob("{}/srunner/configs/*.xosc".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
+        list_of_config_files = glob.glob("{}/srunner/examples/*.xml".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
+        list_of_config_files += glob.glob("{}/srunner/examples/*.xosc".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
 
         if config_file_name != '':
             list_of_config_files.append(config_file_name)
@@ -243,7 +114,7 @@ class ScenarioConfigurationParser(object):
         Parse *all* config files and find first match for scenario config
         """
 
-        list_of_config_files = glob.glob("{}/srunner/configs/*.xml".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
+        list_of_config_files = glob.glob("{}/srunner/examples/*.xml".format(os.getenv('ROOT_SCENARIO_RUNNER', "./")))
 
         if config_file_name != '':
             list_of_config_files.append(config_file_name)
